:bug: Conditionally reconnect audio nodes when recording
When recording, a new audio context destination is created. For most
media layers to be included in the recorded video, its media source node
must be reconnected to the new audio context destination.

Now, only media source nodes that are directly connected to a
destination will be reconnected to the new destination (keeping the
audio graph intact).

Fixes #38.